+++ asdf/tests/test_generic_io.py
@@ -28,6 +28,12 @@ def tree(request):
 
 def _roundtrip(tree, get_write_fd, get_read_fd,
                write_options={}, read_options={}):
+
+    # Since we're testing with small arrays, force all arrays to be stored
+    # in internal blocks rather than letting some of them be automatically put
+    # inline.
+    write_options.setdefault('all_array_storage', 'internal')
+
     with get_write_fd() as fd:
         asdf.AsdfFile(tree).write_to(fd, **write_options)
         # Work around the fact that generic_io's get_file doesn't have a way of